Refactor some group logic out of index.ts
